Bug Fixing

+++ ALERT_CONFIGURATION/General.vb
@@ -318,7 +318,7 @@ and i.id_institution = " & i_ID_INST & "order by 1 asc"
 
     End Function
 
-    Public Function GET_CLIN_SERV(ByVal i_ID_INST As Int16, ByVal i_ID_SOFT As Int16, ByVal i_conn As OracleConnection, ByRef i_dr As OracleDataReader) As Boolean
+    Public Function GET_CLIN_SERV(ByVal i_ID_INST As Int64, ByVal i_ID_SOFT As Int16, ByVal i_conn As OracleConnection, ByRef i_dr As OracleDataReader) As Boolean
 
         Dim l_id_language As Int16 = GET_ID_LANG(i_ID_INST, i_conn)
